+++ src/main/java/roomescape/controller/ReservationController.java
@@ -10,33 +10,33 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.controller.request.ReservationRequest;
-import roomescape.dao.ReservationDAO;
 import roomescape.model.Reservation;
+import roomescape.service.ReservationService;
 
 @RestController
 public class ReservationController {
-    private final ReservationDAO reservationDAO;
 
-    public ReservationController(ReservationDAO reservationDAO) {
-        this.reservationDAO = reservationDAO;
+    private final ReservationService reservationService;
+
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @GetMapping("/reservations")
     public ResponseEntity<List<Reservation>> getReservations() {
-        return ResponseEntity.ok(reservationDAO.findAllReservations());
+        return ResponseEntity.ok(reservationService.getReservations());
     }
 
     @PostMapping("/reservations")
-    public ResponseEntity<Reservation> addReservation(@RequestBody ReservationRequest reservationDto) {
-        Reservation reservation =
-                new Reservation(reservationDto.name(), reservationDto.date(), reservationDto.time());
-        long id = reservationDAO.add(reservation);
+    public ResponseEntity<Reservation> addReservation(@RequestBody ReservationRequest request) {
+        Reservation reservation = new Reservation(request.name(), request.date(), request.time());
+        long id = reservationService.addReservations(reservation);
         return ResponseEntity.created(URI.create("/reservations/" + id)).body(reservation);
     }
 
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> removeReservation(@PathVariable("id") long id) {
-        reservationDAO.delete(id);
+        reservationService.deleteReservation(id);
         return ResponseEntity.noContent().build();
     }
 }
