+++ src/main/java/roomescape/model/Reservation.java
@@ -6,9 +6,11 @@ import java.time.LocalTime;
 public class Reservation {
 
     private long id;
-    private final String name;
-    private final LocalDate date;
-    private final LocalTime time;
+    private String name;
+    private LocalDate date;
+    private LocalTime time;
+
+    public Reservation() {}
 
     public Reservation(long id, String name, LocalDate date, LocalTime time) {
         this.id = id;
