+++ src/main/java/roomescape/repository/ReservationDAO.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.repository;
 
 import java.sql.PreparedStatement;
 import java.util.List;
